Adjusted modal logic for creating todos
